+++ lib/draggable_builder.dart
@@ -2,4 +2,5 @@ library draggable_builder;
 
 export 'src/draggable_builder.dart';
 export 'src/draggable_controller.dart';
+export 'src/draggable_data.dart';
 export 'src/draggable_utils.dart';